bkk-data-pipeline/producer/requirements.txt
kafka-python
requests
schedule
gtfs-realtime-bindings
protobuf